# api-contracts.yaml
...
openapi: 3.0.3
info:
  title: Portfolio Website API
  description: RESTful API for managing portfolio website content
  version: 1.0.0
  contact:
    name: Muhammad Alfarizi Tazkia
    email: muhammadalfarizi.t@gmail.com
servers:
  - url: https://api.portfolio.com/v1
    description: Production server
  - url: http://localhost:8080/v1
    description: Development server
paths:
...
  /skills/{id}:
    put:
      tags: [Skills]
      summary: Update skill
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillInput'
      responses:
        '200':
          description: Skill updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'
    
    delete:
      tags: [Skills]
      summary: Delete skill
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Skill deleted

  # Work Experiences
...
  /work-experiences/{id}:
    get:
      tags: [Work Experiences]
      summary: Get work experience by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Work experience details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkExperience'
    
    put:
      tags: [Work Experiences]
      summary: Update work experience
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkExperienceInput'
      responses:
        '200':
          description: Work experience updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkExperience'
    
    delete:
      tags: [Work Experiences]
      summary: Delete work experience
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Work experience deleted

  # Projects
...
  /projects/{slug}:
    get:
      tags: [Projects]
      summary: Get project by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    
    put:
      tags: [Projects]
      summary: Update project
      security:
        - BearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectInput'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    
    delete:
      tags: [Projects]
      summary: Delete project
      security:
        - BearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Project deleted

  # Articles
...
  /articles/{slug}:
    get:
      tags: [Articles]
      summary: Get article by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Article details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
    
    put:
      tags: [Articles]
      summary: Update article
      security:
        - BearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleInput'
      responses:
        '200':
          description: Article updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
    
    delete:
      tags: [Articles]
      summary: Delete article
      security:
        - BearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Article deleted
...
  /article-categories/{slug}:
    get:
      tags: [Article Categories]
      summary: Get category with articles
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Category with articles
          content:
            application/json:
              schema:
                type: object
                properties:
                  category:
                    $ref: '#/components/schemas/ArticleCategory'
                  articles:
                    type: array
                    items:
                      $ref: '#/components/schemas/ArticleSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

  # Education
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
        created_at:
          type: string
          format: date-time
...
    Skill:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        category:
          type: string
        proficiency_level:
          type: integer
          minimum: 1
          maximum: 5
        is_active:
          type: boolean
        sort_order:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SkillInput:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        category:
          type: string
        proficiency_level:
          type: integer
          minimum: 1
          maximum: 5
        is_active:
          type: boolean
        sort_order:
          type: integer
      required: [name]
    Company:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description_en:
          type: string
        description_ja:
          type: string
        logo_url:
          type: string
        website_url:
          type: string
        industry:
          type: string
        location:
          type: string
    WorkExperience:
      type: object
      properties:
        id:
          type: string
          format: uuid
        company:
          $ref: '#/components/schemas/Company'
        position_en:
          type: string
        position_ja:
          type: string
        employment_type:
          type: string
          enum: [full_time, part_time, contract, internship]
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
          nullable: true
        is_current:
          type: boolean
        description_en:
          type: string
        description_ja:
          type: string
        key_achievements_en:
          type: array
          items:
            type: string
        key_achievements_ja:
          type: array
          items:
            type: string
        detailed_contributions_en:
          type: array
          items:
            type: string
        detailed_contributions_ja:
          type: array
          items:
            type: string
        technologies_used:
          type: array
          items:
            type: string
        sort_order:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    WorkExperienceInput:
      type: object
      properties:
        company_id:
          type: string
          format: uuid
        position_en:
          type: string
        position_ja:
          type: string
        employment_type:
          type: string
          enum: [full_time, part_time, contract, internship]
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
          nullable: true
        is_current:
          type: boolean
        description_en:
          type: string
        description_ja:
          type: string
        key_achievements_en:
          type: array
          items:
            type: string
        key_achievements_ja:
          type: array
          items:
            type: string
        detailed_contributions_en:
          type: array
          items:
            type: string
        detailed_contributions_ja:
          type: array
          items:
            type: string
        technologies_used:
          type: array
          items:
            type: string
        sort_order:
          type: integer
      required: [company_id, position_en, position_ja, start_date]
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        slug:
          type: string
        role_en:
          type: string
        role_ja:
          type: string
        description_en:
          type: string
        description_ja:
          type: string
        technologies_used:
          type: array
          items:
            type: string
        project_url:
          type: string
        github_url:
          type: string
        demo_url:
          type: string
        is_public:
          type: boolean
        is_featured:
          type: boolean
        thumbnail_url:
          type: string
        images:
          type: array
          items:
            type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        status:
          type: string
          enum: [in_progress, completed, on_hold]
        sort_order:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProjectInput:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        role_en:
          type: string
        role_ja:
          type: string
        description_en:
          type: string
        description_ja:
          type: string
        technologies_used:
          type: array
          items:
            type: string
        project_url:
          type: string
        github_url:
          type: string
        demo_url:
          type: string
        is_public:
          type: boolean
        is_featured:
          type: boolean
        thumbnail_url:
          type: string
        images:
          type: array
          items:
            type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        status:
          type: string
          enum: [in_progress, completed, on_hold]
        sort_order:
          type: integer
      required: [title, slug, role_en, role_ja, description_en, description_ja]
    ArticleCategory:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name_en:
          type: string
        name_ja:
          type: string
        slug:
          type: string
        description_en:
          type: string
        description_ja:
          type: string
        icon:
          type: string
        color:
          type: string
        article_count:
          type: integer
        is_active:
          type: boolean
        sort_order:
          type: integer
...
    Article:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title_en:
          type: string
        title_ja:
          type: string
        slug:
          type: string
        description_en:
          type: string
        description_ja:
          type: string
        content_en:
          type: string
        content_ja:
          type: string
        category:
          $ref: '#/components/schemas/ArticleCategory'
        author:
          $ref: '#/components/schemas/User'
        thumbnail_url:
          type: string
        is_published:
          type: boolean
        is_featured:
          type: boolean
        published_at:
          type: string
          format: date-time
        read_time:
          type: integer
        views_count:
          type: integer
        likes_count:
          type: integer
        tags:
          type: array
          items:
            type: string
        meta_title_en:
          type: string
        meta_title_ja:
          type: string
        meta_description_en:
          type: string
        meta_description_ja:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ArticleSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title_en:
          type: string
        title_ja:
          type: string
        slug:
          type: string
        description_en:
          type: string
        description_ja:
          type: string
        category:
          $ref: '#/components/schemas/ArticleCategory'
        author:
          $ref: '#/components/schemas/User'
        thumbnail_url:
          type: string
        published_at:
          type: string
          format: date-time
        read_time:
          type: integer
        views_count:
          type: integer
        likes_count:
          type: integer
        tags:
          type: array
          items:
            type: string
    ArticleInput:
      type: object
      properties:
        title_en:
          type: string
        title_ja:
          type: string
        slug:
          type: string
        description_en:
          type: string
        description_ja:
          type: string
        content_en:
          type: string
        content_ja:
          type: string
        category_id:
          type: string
          format: uuid
        thumbnail_url:
          type: string
        is_published:
          type: boolean
        is_featured:
          type: boolean
        read_time:
          type: integer
        tags:
          type: array
          items:
            type: string
        meta_title_en:
          type: string
        meta_title_ja:
          type: string
        meta_description_en:
          type: string
        meta_description_ja:
          type: string
      required: [title_en, title_ja, slug, description_en, description_ja, content_en, content_ja]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
...